refactor: make web as default implementation
