+++ src/decorators/OpacityDecorator/index.tsx
@@ -1,11 +1,5 @@
-import React from 'react';
-import { OpacityDecorator as OpacityDecoratorDraggableFlatlist } from 'react-native-draggable-flatlist';
 import type { Props } from '../types';
 
 export function OpacityDecorator({ children }: Props) {
-  return (
-    <OpacityDecoratorDraggableFlatlist>
-      {children}
-    </OpacityDecoratorDraggableFlatlist>
-  );
+  return children;
 }
