--- src/decorators/OpacityDecorator/index.web.tsx
@@ -1,5 +1,0 @@
-import type { Props } from '../types';
-
-export function OpacityDecorator({ children }: Props) {
-  return children;
-}
