+++ src/decorators/ScaleDecorator/index.native.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { ScaleDecorator as ScaleDecoratorDraggableFlatlist } from 'react-native-draggable-flatlist';
+import type { Props } from '../types';
+
+export function ScaleDecorator({ children }: Props) {
+  return (
+    <ScaleDecoratorDraggableFlatlist>
+      {children}
+    </ScaleDecoratorDraggableFlatlist>
+  );
+}
