--- src/decorators/ScaleDecorator/index.web.tsx
@@ -1,5 +1,0 @@
-import type { Props } from '../types';
-
-export function ScaleDecorator({ children }: Props) {
-  return children;
-}
