+++ src/decorators/ShadowDecorator/index.tsx
@@ -1,11 +1,5 @@
-import React from 'react';
-import { ShadowDecorator as ShadowDecoratorDraggableFlatlist } from 'react-native-draggable-flatlist';
 import type { Props } from '../types';
 
 export function ShadowDecorator({ children }: Props) {
-  return (
-    <ShadowDecoratorDraggableFlatlist>
-      {children}
-    </ShadowDecoratorDraggableFlatlist>
-  );
+  return children;
 }
