--- src/decorators/ShadowDecorator/index.web.tsx
@@ -1,5 +1,0 @@
-import type { Props } from '../types';
-
-export function ShadowDecorator({ children }: Props) {
-  return children;
-}
