+++ src/index.tsx
@@ -1,8 +1,8 @@
 import { DraggableList } from './DraggableList';
 import 'react-native-gesture-handler';
 
-export { OpacityDecorator } from './decorators/OpacityDecorator';
-export { ScaleDecorator } from './decorators/ScaleDecorator';
-export { ShadowDecorator } from './decorators/ShadowDecorator';
+export { OpacityDecorator } from './decorators/OpacityDecorator/index.native';
+export { ScaleDecorator } from './decorators/ScaleDecorator/index.native';
+export { ShadowDecorator } from './decorators/ShadowDecorator/index.native';
 
 export default DraggableList;
